requirements.txt
langchain
langchain-community
chromadb
sentence-transformers
pypdf
pymupdf
langchain-core
fastapi
uvicorn
streamlit
langchain-google-genai
python-dotenv
